backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
numpy
opencv-python
